backend/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
yfinance>=0.2.54
python-dotenv>=1.0.1
